update, delete auth check. clearing post-user rltn

+++ server/controllers/feed.js
@@ -5,6 +5,7 @@ const { validationResult } = require("express-validator");
 
 const Post = require("../models/post");
 const User = require("../models/user");
+const user = require("../models/user");
 
 exports.getPosts = (req, res, next) => {
   const currentPage = req.query.page || 1;
@@ -15,6 +16,7 @@ exports.getPosts = (req, res, next) => {
     .then((totalDocs) => {
       totalItems = totalDocs;
       return Post.find()
+        .populate("creator", ["name", "email"])
         .skip((currentPage - 1) * perPage)
         .limit(perPage);
     })
@@ -95,6 +97,7 @@ exports.createPost = (req, res, next) => {
 exports.getPost = (req, res, next) => {
   const postId = req.params.postId;
   Post.findById(postId)
+    .populate("creator", "name")
     .then((post) => {
       if (!post) {
         const error = new Error("Post not found.");
@@ -140,6 +143,11 @@ exports.updatePost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error("Not authorized");
+        error.statusCode = 401;
+        throw error;
+      }
       if (imageUrl !== post.imageUrl) {
         clearImage(post.imageUrl);
       }
@@ -171,8 +179,20 @@ exports.deletePost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error("Not authorized");
+        error.statusCode = 401;
+        throw error;
+      }
       clearImage(post.imageUrl);
       return Post.findByIdAndRemove(postId);
+    })
+    .then((result) => {
+      return User.findById(req.userId);
+    })
+    .then((user) => {
+      user.posts.pull(postId);
+      return user.save();
     })
     .then((resutl) => {
       res.status(200).json({ message: "Post Deleted." });
